Scripts/working/ms_coco_open_clip_requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
cmake==3.26.1
filelock==3.10.7
ftfy==6.1.1
huggingface-hub==0.13.3
idna==3.4
Jinja2==3.1.2
lit==16.0.0
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
open-clip-torch==2.16.0
packaging==23.0
Pillow==9.5.0
protobuf==3.20.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
sentencepiece==0.1.97
sympy==1.11.1
timm==0.6.13
tokenizers==0.13.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.27.4
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6